spike: unresolved import

+++ packages/fuel-indexer-database/postgres/src/lib.rs
@@ -632,18 +632,18 @@ pub async fn last_block_height_for_indexer(
     conn: &mut PoolConnection<Postgres>,
     namespace: &str,
     identifier: &str,
-) -> sqlx::Result<u64> {
+) -> sqlx::Result<u32> {
     let query = format!(
         "SELECT MAX(id) FROM {namespace}_{identifier}.indexmetadataentity LIMIT 1"
     );
 
     let row = sqlx::query(&query).fetch_one(conn).await?;
-    let id: i64 = match row.try_get(0) {
+    let id: i32 = match row.try_get(0) {
         Ok(id) => id,
         Err(_e) => return Ok(1),
     };
 
-    Ok(id as u64)
+    Ok(id as u32)
 }
 
 // TODO: https://github.com/FuelLabs/fuel-indexer/issues/251
